requirements.txt
pillow==11.3.0
pandas==2.3.2

